AttachmentClassLoader isn't directly exposed on the api, so move together with serialisation.
