+++ core/src/main/kotlin/net/corda/core/serialization/AttachmentsClassLoader.kt
@@ -1,8 +1,7 @@
-package net.corda.core.node
+package net.corda.core.serialization
 
 import net.corda.core.contracts.Attachment
 import net.corda.core.crypto.SecureHash
-import net.corda.core.serialization.CordaSerializable
 import java.io.ByteArrayInputStream
 import java.io.ByteArrayOutputStream
 import java.io.FileNotFoundException
